pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["craig <73603201+craig-lilley@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "2.2"
numpy = "^1.26.4"
pyarrow = "^15.0.0"
openpyxl = "^3.1.2"
pyqt6 = "^6.6.1"
seaborn = "^0.13.2"
plotly = "^5.19.0"
pyqtgraph = "^0.13.3"
dash = "^2.16.0"
pyzmq = "^25.1.2"
pyqt6-webengine = "^6.6.0"
psutil = "^5.9.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
